init UserModelTestCase: init UserBaseTestCase for refactoring
