+++ english_diary/users/tests/test_api_view.py
@@ -1,37 +1,12 @@
 from django.core.urlresolvers import reverse
-from django.test import TestCase
 from django.contrib.auth import get_user_model
 
-from rest_framework.test import APIClient
 from rest_framework import status
 
-from english_diary.celery import app
+from .base import UserBaseTestCase
 
 
-class UserAPIViewTestCase(TestCase):
-
-    def setUp(self):
-
-        # Run celery task synchronous
-        app.conf.update(CELERY_ALWAYS_EAGER=True)
-
-        self.test_username = "test_username"
-        self.test_password = "test_password"
-        self.test_email = "test@example.com"
-
-        # Create a user
-        self.user = get_user_model().objects.create_user(
-            username=self.test_username,
-            password=self.test_password,
-            email=self.test_email,
-        )
-
-        # Login
-        self.client = APIClient()
-        self.client.login(
-            username=self.test_username,
-            password=self.test_password,
-        )
+class UserAPIViewTestCase(UserBaseTestCase):
 
     def test_user_change_email(self):
